Move filter to http-api
