+++ http-generator-helidon/pom.xml
@@ -23,14 +23,6 @@
       <artifactId>avaje-http-generator-core</artifactId>
       <version>${project.version}</version>
     </dependency>
-
-    <dependency>
-      <groupId>io.avaje</groupId>
-      <artifactId>avaje-prisms</artifactId>
-      <version>${avaje.prisms.version}</version>
-      <scope>provided</scope>
-      <optional>true</optional>
-    </dependency>
   </dependencies>
 
   <build>
@@ -41,13 +33,7 @@
         <version>3.10.1</version>
         <configuration>
           <release>20</release> 
-          <annotationProcessorPaths>
-            <path>
-              <groupId>io.avaje</groupId>
-              <artifactId>avaje-prisms</artifactId>
-              <version>${avaje.prisms.version}</version>
-            </path>
-          </annotationProcessorPaths>
+          <compilerArgument>-proc:none</compilerArgument>
         </configuration>
       </plugin>
     </plugins>
