--- http-generator-helidon/src/main/java/io/avaje/http/generator/helidon/nima/AbstractFilterPrism.java
@@ -1,13 +1,0 @@
-package io.avaje.http.generator.helidon.nima;
-
-import io.avaje.http.generator.core.CustomWebMethod;
-import io.avaje.http.generator.core.WebMethod;
-
-public abstract class AbstractFilterPrism implements CustomWebMethod {
-
-  @Override
-  public WebMethod webMethod() {
-
-    return HelidonWebMethod.FILTER;
-  }
-}
