+++ http-generator-helidon/src/main/java/io/avaje/http/generator/helidon/nima/ControllerWriter.java
@@ -10,6 +10,7 @@ import java.util.Map;
 import io.avaje.http.generator.core.BaseControllerWriter;
 import io.avaje.http.generator.core.Constants;
 import io.avaje.http.generator.core.ControllerReader;
+import io.avaje.http.generator.core.CoreWebMethod;
 import io.avaje.http.generator.core.JsonBUtil;
 import io.avaje.http.generator.core.MethodReader;
 import io.avaje.http.generator.core.PrimitiveUtil;
@@ -51,7 +52,7 @@ class ControllerWriter extends BaseControllerWriter {
     }
     if (reader.methods().stream()
         .map(MethodReader::webMethod)
-        .anyMatch(w -> HelidonWebMethod.FILTER == w)) {
+        .anyMatch(w -> CoreWebMethod.FILTER == w)) {
       reader.addImportType("io.helidon.nima.webserver.http.FilterChain");
       reader.addImportType("io.helidon.nima.webserver.http.RoutingRequest");
       reader.addImportType("io.helidon.nima.webserver.http.RoutingResponse");
