+++ http-generator-helidon/src/main/java/io/avaje/http/generator/helidon/nima/NimaProcessor.java
@@ -5,10 +5,7 @@ import java.io.IOException;
 import io.avaje.http.generator.core.BaseProcessor;
 import io.avaje.http.generator.core.ControllerReader;
 import io.avaje.http.generator.core.PlatformAdapter;
-import io.avaje.http.helidon.Filter;
-import io.avaje.prism.GeneratePrism;
 
-@GeneratePrism(value = Filter.class, superClass = AbstractFilterPrism.class)
 public class NimaProcessor extends BaseProcessor {
 
   @Override
