+++ http-generator-helidon/src/main/java/module-info.java
@@ -1,7 +1,5 @@
 module io.avaje.http.nima.generator {
 
-  exports io.avaje.http.helidon;
-
   requires java.compiler;
   requires java.sql;
 
@@ -9,5 +7,4 @@ module io.avaje.http.nima.generator {
 
   // SHADED: All content after this line will be removed at package time
   requires transitive io.avaje.http.generator.core;
-  requires io.avaje.prism;
 }
