+++ http-generator-javalin/src/main/java/io/avaje/http/generator/javalin/ControllerMethodWriter.java
@@ -2,6 +2,8 @@ package io.avaje.http.generator.javalin;
 
 import static io.avaje.http.generator.core.ProcessingContext.*;
 
+import java.util.List;
+
 import io.avaje.http.generator.core.Append;
 import io.avaje.http.generator.core.CoreWebMethod;
 import io.avaje.http.generator.core.MethodParam;
@@ -25,7 +27,8 @@ class ControllerMethodWriter {
   ControllerMethodWriter(MethodReader method, Append writer, boolean useJsonB) {
     this.method = method;
     this.writer = writer;
-    this.webMethod = method.webMethod();
+    final var webM = method.webMethod();
+    this.webMethod = webM == CoreWebMethod.FILTER ? JavalinWebMethod.BEFORE : webM;
     this.useJsonB = useJsonB && !disabledDirectWrites();
     this.instrumentContext = method.instrumentContext();
     customMethod = !(webMethod instanceof CoreWebMethod);
@@ -35,36 +38,9 @@ class ControllerMethodWriter {
     final var segments = method.pathSegments();
     final var fullPath = segments.fullPath();
 
-    if (method.isErrorMethod()) {
-      writer
-          .append("    app.exception(%s.class, (ex, ctx) -> {", method.exceptionShortName())
-          .eol();
+    writeMethod(fullPath);
 
-    } else {
-      writer.append("    app.%s(\"%s\", ctx -> {", webMethod.name().toLowerCase(), fullPath).eol();
-    }
-
-    if (!customMethod) {
-      writer.append("      ctx.status(%s);", method.statusCode()).eol();
-    }
-
-    final var matrixSegments = segments.matrixSegments();
-    for (final PathSegments.Segment matrixSegment : matrixSegments) {
-      matrixSegment.writeCreateSegment(writer, platform());
-    }
-
-    final var params = method.params();
-    for (final MethodParam param : params) {
-      if (isAssignable2Interface(param.utype().mainType(), "java.lang.Exception")) {
-        continue;
-      }
-      param.writeCtxGet(writer, segments);
-    }
-    if (method.includeValidate()) {
-      for (final MethodParam param : params) {
-        param.writeValidate(writer);
-      }
-    }
+    final var params = writeParams(segments);
     writer.append("      ");
     if (!method.isVoid() && !customMethod) {
       writer.append("var result = ");
@@ -116,6 +92,42 @@ class ControllerMethodWriter {
       }
     }
     writer.append(");").eol().eol();
+  }
+
+  private void writeMethod(final String fullPath) {
+    if (method.isErrorMethod()) {
+      writer
+          .append("    app.exception(%s.class, (ex, ctx) -> {", method.exceptionShortName())
+          .eol();
+
+    } else {
+      writer.append("    app.%s(\"%s\", ctx -> {", webMethod.name().toLowerCase(), fullPath).eol();
+    }
+
+    if (!customMethod) {
+      writer.append("      ctx.status(%s);", method.statusCode()).eol();
+    }
+  }
+
+  private List<MethodParam> writeParams(final PathSegments segments) {
+    final var matrixSegments = segments.matrixSegments();
+    for (final PathSegments.Segment matrixSegment : matrixSegments) {
+      matrixSegment.writeCreateSegment(writer, platform());
+    }
+
+    final var params = method.params();
+    for (final MethodParam param : params) {
+      if (isAssignable2Interface(param.utype().mainType(), "java.lang.Exception")) {
+        continue;
+      }
+      param.writeCtxGet(writer, segments);
+    }
+    if (method.includeValidate()) {
+      for (final MethodParam param : params) {
+        param.writeValidate(writer);
+      }
+    }
+    return params;
   }
 
   private void writeContextReturn() {
