+++ tests/test-nima-jsonb/src/main/java/org/example/TestController.java
@@ -8,6 +8,7 @@ import io.avaje.http.api.BodyString;
 import io.avaje.http.api.Controller;
 import io.avaje.http.api.Default;
 import io.avaje.http.api.ExceptionHandler;
+import io.avaje.http.api.Filter;
 import io.avaje.http.api.Form;
 import io.avaje.http.api.FormParam;
 import io.avaje.http.api.Get;
@@ -16,7 +17,6 @@ import io.avaje.http.api.Path;
 import io.avaje.http.api.Post;
 import io.avaje.http.api.Produces;
 import io.avaje.http.api.QueryParam;
-import io.avaje.http.helidon.Filter;
 import io.helidon.nima.webserver.http.FilterChain;
 import io.helidon.nima.webserver.http.RoutingResponse;
 import io.helidon.nima.webserver.http.ServerRequest;
